% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{compairs_plot}
\alias{compairs_plot}
\title{Pairwise plots of posterior draws}
\usage{
compairs_plot(
  x,
  y = NULL,
  pars = NULL,
  scatter = TRUE,
  overlay = TRUE,
  names = NULL,
  ...
)
}
\arguments{
\item{x}{a moultmcmc object}

\item{y}{an optional second moultmcmc object}

\item{pars}{An optional character vector of parameter names. If pars is not specified then the default is to use the first 4 parameters of x.}

\item{scatter}{logical, draw scatterplot of posterior samples if overlay = FALSE, else draw contours}

\item{overlay}{logical, plot scatterplots and density contours of both models in each triangle, else draw one model per triangle}

\item{...}{further arguments to plot.default (the call that draws the scatter/contour plot)}
}
\description{
Plots pairwise scatter/density plots of posterior draws
}
