% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fixef.moultmcmc}
\alias{fixef.moultmcmc}
\alias{fixef}
\title{Extract Population-Level Estimates}
\usage{
\method{fixef}{moultmcmc}(object, summary = TRUE, probs = c(0.025, 0.975), pars = NULL, ...)
}
\arguments{
\item{pars}{Optional names of coefficients to extract.
By default, all coefficients are extracted.}

\item{...}{Currently ignored.}
}
\value{
If \code{summary} is \code{TRUE}, a matrix returned
by \code{\link{posterior_summary}} for the population-level effects.
If \code{summary} is \code{FALSE}, a matrix with one row per
posterior draw and one column per population-level effect.
}
\description{
Extract the population-level ('fixed') effects
from a \code{moultmcmc} object.
}
\examples{
\dontrun{
fit <- brm(time | cens(censored) ~ age + sex + disease,
           data = kidney, family = "exponential")
fixef(fit)
# extract only some coefficients
fixef(fit, pars = c("age", "sex"))
}

}
