% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{moult_plot.moult}
\alias{moult_plot.moult}
\title{Moult plot for moult model}
\usage{
\method{moult_plot}{moult}(x, prob = 0.95, data = NULL, ...)
}
\arguments{
\item{x}{moult model object created with moult::moult}

\item{prob}{coverage probability of active moult in the sampled population}

\item{data}{optional data.frame of observations, column names need to match model date and moult index variables}

\item{...}{not currently used}
}
\value{
a plot
}
\description{
Moult plot for moult model
}
