% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{moult_plot.moultmcmc}
\alias{moult_plot.moultmcmc}
\title{Moult plot for moultmcmc model}
\usage{
\method{moult_plot}{moultmcmc}(
  x,
  prob = 0.95,
  prob_ci = NULL,
  plot_data = TRUE,
  plot = TRUE,
  newdata = NULL,
  col.data = NULL,
  ...
)
}
\arguments{
\item{x}{moultmcmc model object}

\item{prob}{coverage probability of active moult in the sampled population}

\item{prob_ci}{coverage probability of credible interval}

\item{plot_data}{logical, if TRUE plot observations}

\item{plot}{logical, if TRUE (default) return a plot, else return a dataframe with start and end dates of moult based on model intercepts}

\item{newdata}{optional data.frame of conditions to plot (passed to predict.moultmcmc)}

\item{col.data}{optional, name of column in model data to colour data points}

\item{...}{not currently used}
}
\value{
a plot or data.frame
}
\description{
Moult plot for moultmcmc model
}
