% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moultmcmc.R
\name{moultmcmc}
\alias{moultmcmc}
\title{Bayesian inference for Underhill-Zucchini moult models and expansions}
\usage{
moultmcmc(
  moult_column,
  date_column,
  id_column = NULL,
  start_formula = ~1,
  duration_formula = ~1,
  sigma_formula = ~1,
  type = 2,
  lump_non_moult = FALSE,
  data,
  init = "auto",
  flat_prior = TRUE,
  beta_sd = 0,
  log_lik = FALSE,
  use_phi_approx = FALSE,
  active_moult_recaps_only = FALSE,
  ...
)
}
\arguments{
\item{moult_column}{the name the column in \code{data} containing moult indices, i.e. a numeric vector of (linearized) moult scores (0 = old plumage,1 = new plumage) or numerical moult codes (1 = old plumage,2 = moulting, 3 = new plumage), depending on the model type.}

\item{date_column}{the name the column in \code{data} containing sampling dates, encoded as days since an arbitrary reference date, i.e. a numeric vector}

\item{id_column}{(optional) factor identifier. Usually a season-individual combination to encode within-season recaptures, defaults to NULL. When provided moultmcmc will attempt to fit the relevant recaptures model.}

\item{start_formula}{model formula for start date}

\item{duration_formula}{model formula for duration}

\item{sigma_formula}{model formula for start date sigma}

\item{type}{integer (one of 1,2,3,4,5,12) referring to type of moult data and consequently model to be fitted (see details)}

\item{lump_non_moult}{logical; should pre- and post-moult observations be treated as indistinguishable? if TRUE and type \%in\% c(1,2,12), the relevant lumped model will be fitted (see details).}

\item{data}{Input data frame must contain a numeric column "date" and a column "moult_cat" which is a numeric vector of categorical moult codes (1 = old plumage,2 = moulting,3 = new plumage).}

\item{init}{Specification of initial values for all or some parameters. Can be the string "auto" for an automatic guess based on the data, or any of the permitted rstan options: the digit 0, the strings "0" or "random", or a function. See the detailed documentation for the init argument in ?rstan::stan.}

\item{flat_prior}{use uniform prior on start date and duration (TRUE) or vaguely informative truncated normal prior (FALSE). Defaults to TRUE.}

\item{beta_sd}{use zero-centred normal priors for regression coefficients other than intercepts? If <= 0 the stan default of improper flat priors is used.}

\item{log_lik}{boolean retain pointwise log-likelihood in output? This enables model assessment and selection via the loo package. Defaults to FALSE, can lead to very large output arrays when sample size is large.}

\item{use_phi_approx}{logical flag whether to use stan's Phi_approx function to calculate the "old" likelihoods}

\item{active_moult_recaps_only}{logical flag whether to ignore repeated observations outside the active moult phase}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference for Underhill-Zucchini moult models and expansions
}
\details{
type refers to the type of moult data available (see Underhill and Zucchini (1998) and Underhill, Zucchini and Summers (1990)).

type = 1
sample is representative of entire population (not yet moulting, in moult, and birds which have completed moult). For type 1 data, any value between 0 and 1 (> 0 and < 1) can be used as the moult index for birds in active moult. The value used does not matter, only the fact that they are in moult.
type = 2
(default) sample is representative of entire population (not yet moulting, in moult, and birds which have completed moult). Moult scores are required.

type = 3
sample is representative only of birds in moult. Individuals with moult scores 0 or 1 are ignored.

type = 4
sample is representative only of birds in moult and those that have completed moult. Individuals with moult scores 0 are ignored.

type = 5
sample is representative only of birds that have not started moult or that are in moult. Individuals with moult scores 1 are ignored.
}
