% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{namedList}
\alias{namedList}
\title{internal helper function to create a named list}
\usage{
namedList(...)
}
\arguments{
\item{...}{args}
}
\value{
a named list
}
\description{
original version by Ben Bolker: https://stackoverflow.com/a/16951524
}
