% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot.moult}
\alias{plot.moult}
\title{Plot method for moult models}
\usage{
\method{plot}{moult}(x, prob = 0.95, scales = "free_x")
}
\arguments{
\item{x}{maximum likelihood fit of a moult model from the moult package (class moult)}

\item{prob}{coverage probability of the confidence interval}

\item{scales}{argument to facet_wrap. default is "free_x", "free" may enable a better comparison if parameter values are disparate}
}
\value{
a plot
}
\description{
Plot method for moult models
}
