% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.moultmcmc}
\alias{predict.moultmcmc}
\title{Predict method for moultmcmc models}
\usage{
\method{predict}{moultmcmc}(object, newdata = NULL, predict.type = "start", intervals = 0.1, ...)
}
\arguments{
\item{object}{a fitted moultmcmc model}

\item{newdata}{dataframe with explanatory variables for which to make predictions}

\item{predict.type}{specifies form of predictions, see details.}

\item{intervals}{not currently used}

\item{...}{}
}
\value{
a data.frame?!
}
\description{
Predict method for moultmcmc models
}
