% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ranef.moultmcmc}
\alias{ranef.moultmcmc}
\alias{ranef}
\title{Extract Individual-Level Estimates}
\usage{
\method{ranef}{moultmcmc}(object, summary = TRUE, probs = c(0.025, 0.975), pars = NULL, ...)
}
\arguments{
\item{object}{a moultmcmc model}

\item{summary}{logical, should posterior samples be summarised for each parameter}

\item{probs}{numeric, desired quantiles for summary statistics}

\item{pars}{Optional names of coefficients to extract.
By default, all coefficients are extracted.}

\item{...}{Currently ignored.}
}
\value{
If \code{summary} is \code{TRUE}, a matrix for the individual-level effects.
If \code{summary} is \code{FALSE}, a matrix with one row per
posterior draw and one column per individual-level effect.
}
\description{
Extract the individual-level ('random') effects
from a \code{moultmcmc} object.
}
