% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{residual_plot.moult}
\alias{residual_plot.moult}
\title{Residual plot for moult model}
\usage{
\method{residual_plot}{moult}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{moult model object created with moult::moult}

\item{plot}{logical, if TRUE (default) return a plot, else return a data.frame with calculated quantities}

\item{...}{not currently used}
}
\value{
a plot
}
\description{
This function displays a plot that shows where observed dates and moult scores fall relative to the predictions of the fitted model. Active moult observations should generally fall within +/- 3 start-date standard deviations of the regression line.
}
