% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siskins}
\alias{siskins}
\title{Moult records of Eurasian Siskins}
\format{
A data frame with 299 rows and 13 variables:
\describe{
\item{yday}{sampling date (days since Jan 01)}
\item{pfmg}{PFMG at sampling occasion}
\item{id}{individual identifier}
}
}
\source{
Hugh Insley
}
\usage{
siskins
}
\description{
A dataset containing the moult scores and capture dates for Scottish population of Eurasian Siskins (Spinus spinus) including recaptures.
}
\references{
Insley et al. (in prep). Breeding and moult phenology of siskins in the Scottish Highlands.
}
\keyword{datasets}
