% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary.moultmcmc}
\alias{summary.moultmcmc}
\title{Summary method for moultmcmc models}
\usage{
\method{summary}{moultmcmc}(object, ...)
}
\arguments{
\item{object}{a moultmcmc object}

\item{...}{passed to summary_table}
}
\value{
a summary table
}
\description{
Summary method for moultmcmc models
}
