% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summary_table.moultmcmc}
\alias{summary_table.moultmcmc}
\title{Summary Table}
\usage{
\method{summary_table}{moultmcmc}(
  x,
  pars = x$stanfit@sim$pars_oi,
  prob = 0.95,
  include = TRUE,
  ...
)
}
\arguments{
\item{x}{moultmcmc fit object}

\item{pars}{A character vector of parameter names. The default is all parameters for which samples are saved. If include = FALSE, then the specified parameters are excluded from the printed summary.}

\item{prob}{nominal coverage probability of credible interval}

\item{include}{Logical scalar (defaulting to TRUE) indicating whether to include or exclude the parameters named by the pars argument.}

\item{...}{Additional arguments passed to the summary method for stanfit objects.}
}
\value{

}
\description{
Summary Table
}
