% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz1_linpred.R
\name{uz1_linpred}
\alias{uz1_linpred}
\title{Bayesian inference for the Type 1 moult model}
\usage{
uz1_linpred(
  start_formula = ~1,
  duration_formula = ~1,
  sigma_formula = ~1,
  data,
  init = "auto",
  ...
)
}
\arguments{
\item{start_formula}{model formula for start date}

\item{duration_formula}{model formula for duration}

\item{sigma_formula}{model formula for start date sigma}

\item{data}{Input data frame must contain a numeric column "date" and a column "moult_cat" which is a numeric vector of categorical moult codes (1 = old plumage,2 = moulting,3 = new plumage).}

\item{init}{Specification of initial values for all or some parameters. Can be the string "auto" for an automatic guess based on the data, or any of the permitted rstan options: the digit 0, the strings "0" or "random", or a function. See the detailed documentation for the init argument in ?rstan::stan.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference for the Type 1 moult model
}
