% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz1_simple.R
\name{uz1_simple}
\alias{uz1_simple}
\title{Bayesian inference for the Type 1 moult model}
\usage{
uz1_simple(dates, moult_cat, init = "auto", ...)
}
\arguments{
\item{dates}{Numeric vector of input capture dates.}

\item{moult_cat}{Numeric vector of categorical moult codes (1 = old plumage,2 = moulting,3 = new plumage).
#' @param init Specification of initial values for all or some parameters. Can be the string "auto" for an automatic guess based on the data, or any of the permitted rstan options: the digit 0, the strings "0" or "random", or a function. See the detailed documentation for the init argument in ?rstan::stan.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference for the Type 1 moult model
}
