% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz2_recap.R
\name{uz2_linpred_recap}
\alias{uz2_linpred_recap}
\title{Bayesian inference for the Type 2 moult model with repeat measures}
\usage{
uz2_linpred_recap(
  moult_index_column,
  date_column,
  id_column,
  start_formula = ~1,
  duration_formula = ~1,
  sigma_formula = ~1,
  lump_non_moult = FALSE,
  active_moult_recaps_only = FALSE,
  data,
  init = "auto",
  flat_prior = TRUE,
  beta_sd = 0,
  log_lik = FALSE,
  use_phi_approx = FALSE,
  same_sigma = FALSE,
  ...
)
}
\arguments{
\item{moult_index_column}{the name the column in \code{data} containing moult indices, i.e. a numeric vector of (linearized) moult scores (0 = old plumage,1 = new plumage).}

\item{date_column}{the name the column in \code{data} containing sampling dates, encoded as days since an arbitrary reference date, i.e. a numeric vector}

\item{id_column}{factor identifier. Usually a season-individual combination to encode within-season recaptures}

\item{start_formula}{model formula for start date}

\item{duration_formula}{model formula for duration}

\item{sigma_formula}{model formula for start date sigma}

\item{lump_non_moult}{logical; should pre- and post-moult observations be treated as indistinguishable? if TRUE, the type 2L model will be fitted.}

\item{active_moult_recaps_only}{logical flag whether to ignore repeated observations outside the active moult phase}

\item{data}{Input data frame}

\item{init}{Specification of initial values for all or some parameters. Can be the string "auto" for an automatic guess based on the data, or any of the permitted rstan options: the digit 0, the strings "0" or "random", or a function. See the detailed documentation for the init argument in ?rstan::stan.}

\item{flat_prior}{use uniform prior on start date and duration (TRUE) or vaguely informative truncated normal prior (FALSE). Defaults to TRUE.}

\item{beta_sd}{use zero-centred normal priors for regression coefficients other than intercepts? If <= 0 the stan default of improper flat priors is used.}

\item{log_lik}{boolean retain pointwise log-likelihood in output? This enables model assessment and selection via the loo package. Defaults to FALSE, can lead to very large output arrays if sample size is large.}

\item{use_phi_approx}{logical flag whether to use stan's Phi_approx function to calculate the "old" likelihoods}

\item{same_sigma}{logical flag, currently unused}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference for the Type 2 moult model with repeat measures
}
