% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz5_recap_annual_raneff.R
\name{uz5_linpred_recap_annual_raneff}
\alias{uz5_linpred_recap_annual_raneff}
\title{Bayesian inference for the Type 5 moult model with repeat measures and an additional annual random intercept}
\usage{
uz5_linpred_recap_annual_raneff(
  moult_index_column,
  date_column,
  id_column,
  start_formula = ~1,
  duration_formula = ~1,
  sigma_formula = ~1,
  year_factor_column,
  beta_sd = 0,
  data,
  init = "auto",
  log_lik = TRUE,
  ...
)
}
\arguments{
\item{moult_index_column}{the name the column in \code{data} containing moult indices, i.e. a numeric vector of (linearized) moult scores (0 = old plumage,1 = new plumage).}

\item{date_column}{the name the column in \code{data} containing sampling dates, encoded as days since an arbitrary reference date, i.e. a numeric vector}

\item{id_column}{factor identifier. Usually a season-individual combination to encode within-season recaptures}

\item{start_formula}{model formula for start date}

\item{duration_formula}{model formula for duration}

\item{sigma_formula}{model formula for start date sigma}

\item{year_factor_column}{the name of a factor column in data that contains year_factor values}

\item{beta_sd}{use zero-centred normal priors for regression coefficients other than intercepts? If <= 0 the stan default of improper flat priors is used.}

\item{data}{Input data frame}

\item{init}{Specification of initial values for all or some parameters. Can be the string "auto" for an automatic guess based on the data, or any of the permitted rstan options: the digit 0, the strings "0" or "random", or a function. See the detailed documentation for the init argument in ?rstan::stan.}

\item{log_lik}{boolean retain pointwise log-likelihood in output? This enables model assessment and selection via the loo package. Defaults to true, can lead to very large output arrays if sample size is large.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference for the Type 5 moult model with repeat measures and an additional annual random intercept
}
