% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weavers_processed}
\alias{weavers_processed}
\title{Moult records of Southern Masked Weavers}
\format{
A data frame with 5282 rows and 6 variables:
\describe{
\item{RDate}{sampling date (calendar date)}
\item{Sex}{SAFRING sex code}
\item{Year}{Sampling Year}
\item{Moult}{feather scores for primaries}
\item{pfmg}{Proportion of feather mass grown}
\item{day}{sampling date (days since August 01 of each year)}
}
}
\source{
Erni et al. 2013
}
\usage{
weavers_processed
}
\description{
A dataset containing the moult scores and capture dates for Southern Masked Weavers (Ploceus velatus) from the Western Cape, South Africa.
}
\references{
Oschadleus 2005, Patterns of primary moult in the weavers, Ploceidae. University of Cape Town.
}
\keyword{datasets}
