% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compare_plot}
\alias{compare_plot}
\title{Visual comparison of moult models}
\usage{
compare_plot(..., names = NULL)
}
\arguments{
\item{...}{two or a moult or moultmcmc model}

\item{names}{optional character vector of model names}
}
\value{
a plot comparing parameter estimates and their uncertainties
}
\description{
Visual comparison of moult models
}
