% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compare_plot}
\alias{compare_plot}
\title{Visual comparison of ML and MCMC fits}
\usage{
compare_plot(m1, m2, names = NULL)
}
\arguments{
\item{ml}{a moult model}

\item{mcmc}{a moultmcmc model}
}
\value{
a plot
}
\description{
Visual comparison of ML and MCMC fits
}
