% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{moult_plot.moult}
\alias{moult_plot.moult}
\title{Moult plot for moult model}
\usage{
\method{moult_plot}{moult}(x, prob = 0.95, plot_data = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{moult model object created with moult::moult}

\item{prob}{coverage probability of active moult in the sampled population}

\item{plot_data}{logical, if TRUE plot observations}

\item{plot}{logical, if TRUE (default) return a plot, else return a dataframe with start and end dates of moult based on model intercepts}

\item{...}{not currently used}
}
\value{
a plot or data.frame
}
\description{
Moult plot for moult model
}
