% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{moult_plot.moultmcmc}
\alias{moult_plot.moultmcmc}
\title{Moult plot for moultmcmc model}
\usage{
\method{moult_plot}{moultmcmc}(x, prob = 0.95, prob_ci = NULL, data = NULL, ...)
}
\arguments{
\item{x}{moultmcmc model object}

\item{prob}{coverage probability of active moult in the sampled population}

\item{prob_ci}{coverage probability of credible interval}

\item{data}{optional data.frame of observations, column names need to match model date and moult index variables}

\item{...}{not currently used}
}
\value{
a plot - currently only the intercept terms are taken into account
}
\description{
Moult plot for moultmcmc model
}
