% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siskins}
\alias{siskins}
\title{Moult records of Eurasian Siskins}
\format{
A data frame with 299 rows and 3 variables:
\describe{
\item{yday}{sampling date (days since Jan 01)}
\item{pfmg}{PFMG at sampling occasion}
\item{id}{individual identifier}
\item{start_date} "true" individual start date
\item{duration} "true" individual moult duration
}
}
\source{
Philipp Boersch-Supan
}
\usage{
siskins
}
\description{
A dataset containing simulated moult scores and capture dates of Eurasian Siskins (Spinus spinus), based on a sub-urban Scottish population.
}
\references{
Insley et al. (in prep). Breeding and moult phenology of siskins in the Scottish Highlands.
}
\keyword{datasets}
