% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz1_linpred.R
\name{uz1_linpred}
\alias{uz1_linpred}
\title{Bayesian inference for the Type 1 moult model}
\usage{
uz1_linpred(
  moult_cat_column,
  date_column,
  start_formula = ~1,
  duration_formula = ~1,
  sigma_formula = ~1,
  data,
  init = "auto",
  log_lik = TRUE,
  ...
)
}
\arguments{
\item{moult_cat_column}{the name the column in \code{data} containing moult categories, i.e. a numeric vector of categorical moult codes (1 = old plumage,2 = moulting,3 = new plumage)}

\item{date_column}{the name the column in \code{data} containing sampling dates, encoded as days since an arbitrary reference date, i.e. a numeric vector}

\item{start_formula}{model formula for start date}

\item{duration_formula}{model formula for duration}

\item{sigma_formula}{model formula for start date sigma}

\item{data}{Input data frame}

\item{init}{Specification of initial values for all or some parameters. Can be the string "auto" for an automatic guess based on the data, or any of the permitted rstan options: the digit 0, the strings "0" or "random", or a function. See the detailed documentation for the init argument in ?rstan::stan.}

\item{log_lik}{boolean retain pointwise log-likelihood in output? This enables model assessment and selection via the loo package. Defaults to true, can lead to very large output arrays if sample size is large.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference for the Type 1 moult model
}
