% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uz1_simple.R
\name{uz1_simple}
\alias{uz1_simple}
\title{Bayesian inference for the Type 1 moult model}
\usage{
uz1_simple(dates, moult_cat, guess_inits = TRUE, ...)
}
\arguments{
\item{dates}{Numeric vector of input capture dates.}

\item{moult_cat}{Numeric vector of categorical moult codes (1 = old plumage,2 = moulting,3 = new plumage).}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian inference for the Type 1 moult model
}
